Refactor SemesterPlanGrid for clearer code
